SW-7817 Use Observation Summaries for site survival rate (#4865)
The Zone survival rates on the plants dashboard uses the observation summaries endpoint, but the site survival rate was using the regular results endpoint.
In some cases, the summaries will have values but the results will not.
Update the site survival rate to use the summaries.

Also fix the message at the top of the dashboard to use the summaries survival rate.