SW-7850 Fix planting site update calls (#4887)
Updates to planting sites were happening in 2 places - the planting sites page and the survival rate settings page (which only modified the `survivalRateIncludesTempPlots` flag). Both of these weren't sending the full details of the planting site for the `put` request which caused them both to modify the planting site unintentionally. 
The server PR https://github.com/terraware/terraware-server/pull/3724 contains the flag in the `get` payload for the site which is necessary to fix the planting site page. Update `EditTemporaryPlotsTab` to use rtk query for retrieval and update of the planting site.

A later PR will update several of these pages to use rtk query for the rest and remove the existing redux code once it's no longer used.