SW-7835 Remove survival rate feature flag (#4876)
Note that variables and files that reference mortality rate that include survival rate values will be removed in a later PR.

Note that this should not be merged until QA finalizes.

Note also that parts of the tests were just commented out because they referenced mortality rates, and data has not yet been added to be able to calculate survival rates (which will be part of a separate PR).