Migrate survival rate settings pages to RTK query (#4897)
This isn't quite 100% of queries yet. Later PRs will include the search api requests to rtk query, as well as the t0 all set check on other pages. 

Will likely wait to merge this till after we see if survival rates is going out Thursday, just to make sure we're not causing anything else that could delay the release of survival rates.