SW-7895 Replace references to plantingZoneNames (#4938)
The search api added the new column name of `substratumNames` to the `NurseryWithdrawalsTable` (keeping `plantingSubzoneNames` for backwards compatibility). Replace search api references with the updated column name so it can be removed from the BE. 
Because there weren't many other references to `plantingSubzoneNames`, go ahead and replace the rest of the variables/references with `substratumNames`.